Features to facilitate a new OAuth2 example app

+++ tests/test_client.c
@@ -120,7 +120,7 @@ int test_client(void) {
     return 1;
   }
   if (parsed_count != 2) {
-    printf("Parsed count mismatch: %zu\n", parsed_count);
+    printf("Parsed count mismatch: %lu\n", (unsigned long)parsed_count);
     return 1;
   }
   if (strcmp(parsed_fields[0].key, "grant_type") != 0 ||
